pyproject.toml
[project]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13"
]
license = { file = "LICENSE" }
name = "fusionbrain-sdk-python"
version = "0.1.2"
description = "SDK for FusionBrainAI API"
readme = "README.md"
authors = [
    { name = "Anatoly Korobov", email = "fusionbrain26@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.12.11",
    "async-timeout>=5.0.1",
    "clear>=2.0.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aioresponses>=0.7.8",
    "mypy>=1.16.0",
    "polyfactory>=2.21.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-env>=1.1.5",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.7.0",
    "requests-mock>=1.12.1",
    "ruff>=0.11.12",
    "types-requests>=2.32.0.20250602",
]
